Add `.ruby-version` file
If you're using `rbenv` to manage your Ruby version when working on this
project rather than `rvm`, it uses your global default rather than the
version specified in the Gemfile.

This adds a `.ruby-version` file to the root of the project with a
version matching the Gemfile. It also updates the Gemfile to read from
this file to ensure consistency.

Whether rbenv or rvm is better really doesn't matter – supporting both
is A Good Thing™.