Add support for a "directory" of users
This adds a new endpoint at `/users` that returns a list of all the
users in Screenhole. I've got another PR for jake/screenhole-web that
implements a Yellow Pages-style UI.

I was kind of intrigued to see who's actually using the thing and
thought this might be a fun thing.

As part of this, I've added a `visible_in_directory` scope to the `User`
model. Right now it just returns everyone, but we can use it to filter
out people for whatever reason further down the line.

P.S. MiniTest suuuuuuuuuucks